FIXME: Quick'n'dirty support for files not on the FS
There is a builtin assumption in TeXpresso that files that are being
edited are stored on the filesystem, and that the "changes" overlay the
actual content.
But it can be the case that a file exists only in the editor (it has not
been saved yet).

This fix is a hack: when that's the case, a new TeX process starts from
scratch, and TeXpresso substitutes the file for the virtual one when it
is opened for the first time from the new process.

A proper fix require a redesign of the "trace" subsystem.
The trace should account for every observation of the filesystem.
Also, a trace entry should keep a reference to the previous trace entry
from the same file: this would simplify the rollback and allow detecting
the first event referencing a given file (the first event is the one
without a previous trace entry :)).
Benefits:
- more accurate simulation of the FS semantics
- long term: possibilities to garbage collect old entries
- simpler rollback code (in particular, the invariant in the current
  implementation are difficult to keep track off!!)