Fix incorrect parsing of indirect references
The PDF lexer would tokenize "0 0 0 RG" as "0" "0 0 R" "G" instead of
"0" "0" "0" "RG" because of greedy matching and indirect reference
notation.

This is fixed by requiring a space after the "R". This is a temporary
solution and something cleaner would be beneficial in the long-run.

This should fix issue https://github.com/let-def/texpresso/issues/97